Support `copy_dir_all` in unit tests too
Refactoring for https://github.com/astral-sh/uv/pull/9091
